requirements.txt
lxml_html_clean==0.4.1
newspaper4k==0.9.3.1
nltk==3.9.1
scrapy==2.11.0
mkdocs==1.6.1
